Fixed an issue with redoing from the start of the history state.
It seems like tab.levelHistory.currentPosition (which would be at -1) was silently getting cast to a size_t and making the first condition fail and returning from the function without redoing. This never occurred before so I'm not sure why it is now -- maybe due to the recent MSVC compiler updates/changes?