Marked the current source as old.
We are working on a massive overhaul. I'm not sure of the best way to go about it, but I feel it would make most sense to just build from the ground up and take bits of the old source as needed to integrate with the new architecture. We will see how this method goes.